refactor: break circular dependency loop by moving ToolCall interfaces to scheduler/types.ts
