+++ packages/core/src/core/turn.ts
@@ -5,7 +5,6 @@
  */
 
 import type {
-  Part,
   PartListUnion,
   GenerateContentResponse,
   FunctionCall,
@@ -16,9 +15,7 @@ import type {
 import type {
   ToolCallConfirmationDetails,
   ToolResult,
-  ToolResultDisplay,
 } from '../tools/tools.js';
-import type { ToolErrorType } from '../tools/tool-error.js';
 import { getResponseText } from '../utils/partUtils.js';
 import { reportError } from '../utils/errorReporting.js';
 import {
@@ -33,7 +30,11 @@ import { createUserContent } from '@google/genai';
 import type { ModelConfigKey } from '../services/modelConfigService.js';
 import { getCitations } from '../utils/generateContentResponseUtilities.js';
 
-// Define a structure for tools passed to the server
+import {
+  type ToolCallRequestInfo,
+  type ToolCallResponseInfo,
+} from '../scheduler/types.js';
+
 export interface ServerTool {
   name: string;
   schema: FunctionDeclaration;
@@ -100,26 +101,6 @@ export interface GeminiErrorEventValue {
 export interface GeminiFinishedEventValue {
   reason: FinishReason | undefined;
   usageMetadata: GenerateContentResponseUsageMetadata | undefined;
-}
-
-export interface ToolCallRequestInfo {
-  callId: string;
-  name: string;
-  args: Record<string, unknown>;
-  isClientInitiated: boolean;
-  prompt_id: string;
-  checkpoint?: string;
-  traceId?: string;
-}
-
-export interface ToolCallResponseInfo {
-  callId: string;
-  responseParts: Part[];
-  resultDisplay: ToolResultDisplay | undefined;
-  error: Error | undefined;
-  errorType: ToolErrorType | undefined;
-  outputFile?: string | undefined;
-  contentLength?: number;
 }
 
 export interface ServerToolCallConfirmationDetails {
