+++ packages/core/src/telemetry/loggers.test.circular.ts
@@ -13,10 +13,10 @@ import { logToolCall } from './loggers.js';
 import { ToolCallEvent } from './types.js';
 import type { Config } from '../config/config.js';
 import type { CompletedToolCall } from '../core/coreToolScheduler.js';
-import type {
-  ToolCallRequestInfo,
-  ToolCallResponseInfo,
-} from '../core/turn.js';
+import {
+  type ToolCallRequestInfo,
+  type ToolCallResponseInfo,
+} from '../scheduler/types.js';
 import { MockTool } from '../test-utils/mock-tool.js';
 
 describe('Circular Reference Handling', () => {
